+++ CHANGELOG.md
@@ -8,3 +8,15 @@ Fix release
 ### Fixed
 
 - Fix `username` generator
+
+## [0.0.3] - 2023-05-04
+
+Fix release
+
+### Added
+
+- Support `religion` generator
+
+### Fixed
+
+- Fix `unidecode` import